requirement.txt
asgiref==0.14.0
autobahn==0.16.0
channels==0.15.1
daphne==0.13.1
Django==1.10.1
django-authtools==1.5.0
django-crispy-forms==1.6.0
django-datatable-view==0.8.3
Genshi==0.7
haikunator==1.0.1
Pillow==3.3.1
python-dateutil==2.5.3
six==1.10.0
Twisted==16.2.0
txaio==2.5.1
zope.interface==4.3.2
